requirements.txt
discord.py
aiohttp
python-dotenv
aiosqlite